- Addressed a few static analysis warnings in CacheableScriptFile::OpenBase and CAccount::SetNewPassword. - Added new PROFILE_TYPE, "PROFILE_STAT_FAULTS". This is used to record the number of exceptions that occur on a per-thread basis. This statistic is incremented in every try-catch handler. - Moved EnableProfile(PROFILE_IDLE) and EnableProfile(PROFILE_OVERHEAD) to AbstractSphereThread, since these statistics are always recorded for every thread. - Fixed missing return false in CItemBase::GetItemData when an exception is thrown, would probably have caused an unhandled exception to be thrown due to pData not being initialised.

+++ 0.56.b/common/CacheableScriptFile.cpp
@@ -33,7 +33,7 @@ bool CacheableScriptFile::OpenBase(void *pExtra)
 	m_hFile = (OSFILE_TYPE)STDFUNC_FILENO(m_pStream);
 	m_closed = false;
 	TemporaryString buf;
-	int nStrLen;
+	size_t nStrLen;
 	bool bUTF = false, bFirstLine = true;
 	
 	while( !feof(m_pStream) )
